initial browser refactor to remove additional process
