+++ openhands/runtime/browser/utils.py
@@ -1,60 +1,56 @@
-import os
+import base64
+import io
 
-from openhands.core.exceptions import BrowserUnavailableException
-from openhands.core.schema import ActionType
-from openhands.events.action import BrowseInteractiveAction, BrowseURLAction
-from openhands.events.observation import BrowserOutputObservation
-from openhands.runtime.browser.browser_env import BrowserEnv
+import html2text
+import numpy as np
+from PIL import Image
 
 
-async def browse(
-    action: BrowseURLAction | BrowseInteractiveAction, browser: BrowserEnv | None
-) -> BrowserOutputObservation:
-    if browser is None:
-        raise BrowserUnavailableException()
+def get_html_text_converter():
+    html_text_converter = html2text.HTML2Text()
+    # ignore links and images
+    html_text_converter.ignore_links = False
+    html_text_converter.ignore_images = True
+    # use alt text for images
+    html_text_converter.images_to_alt = True
+    # disable auto text wrapping
+    html_text_converter.body_width = 0
+    return html_text_converter
 
-    if isinstance(action, BrowseURLAction):
-        # legacy BrowseURLAction
-        asked_url = action.url
-        if not asked_url.startswith('http'):
-            asked_url = os.path.abspath(os.curdir) + action.url
-        action_str = f'goto("{asked_url}")'
 
-    elif isinstance(action, BrowseInteractiveAction):
-        # new BrowseInteractiveAction, supports full featured BrowserGym actions
-        # action in BrowserGym: see https://github.com/ServiceNow/BrowserGym/blob/main/core/src/browsergym/core/action/functions.py
-        action_str = action.browser_actions
-    else:
-        raise ValueError(f'Invalid action type: {action.action}')
+def image_to_png_base64_url(
+    image: np.ndarray | Image.Image, add_data_prefix: bool = False
+):
+    """Convert a numpy array to a base64 encoded png image url."""
+    if isinstance(image, np.ndarray):
+        image = Image.fromarray(image)
+    if image.mode in ('RGBA', 'LA'):
+        image = image.convert('RGB')
+    buffered = io.BytesIO()
+    image.save(buffered, format='PNG')
 
-    try:
-        # obs provided by BrowserGym: see https://github.com/ServiceNow/BrowserGym/blob/main/core/src/browsergym/core/env.py#L396
-        obs = browser.step(action_str)
-        return BrowserOutputObservation(
-            content=obs['text_content'],  # text content of the page
-            url=obs.get('url', ''),  # URL of the page
-            screenshot=obs.get('screenshot', None),  # base64-encoded screenshot, png
-            open_pages_urls=obs.get('open_pages_urls', []),  # list of open pages
-            active_page_index=obs.get(
-                'active_page_index', -1
-            ),  # index of the active page
-            dom_object=obs.get('dom_object', {}),  # DOM object
-            axtree_object=obs.get('axtree_object', {}),  # accessibility tree object
-            extra_element_properties=obs.get('extra_element_properties', {}),
-            focused_element_bid=obs.get(
-                'focused_element_bid', None
-            ),  # focused element bid
-            last_browser_action=obs.get(
-                'last_action', ''
-            ),  # last browser env action performed
-            last_browser_action_error=obs.get('last_action_error', ''),
-            error=True if obs.get('last_action_error', '') else False,  # error flag
-        )
-    except Exception as e:
-        return BrowserOutputObservation(
-            content=str(e),
-            screenshot='',
-            error=True,
-            last_browser_action_error=str(e),
-            url=asked_url if action.action == ActionType.BROWSE else '',
-        )
+    image_base64 = base64.b64encode(buffered.getvalue()).decode()
+    return (
+        f'data:image/png;base64,{image_base64}'
+        if add_data_prefix
+        else f'{image_base64}'
+    )
+
+
+def image_to_jpg_base64_url(
+    image: np.ndarray | Image.Image, add_data_prefix: bool = False
+):
+    """Convert a numpy array to a base64 encoded jpeg image url."""
+    if isinstance(image, np.ndarray):
+        image = Image.fromarray(image)
+    if image.mode in ('RGBA', 'LA'):
+        image = image.convert('RGB')
+    buffered = io.BytesIO()
+    image.save(buffered, format='JPEG')
+
+    image_base64 = base64.b64encode(buffered.getvalue()).decode()
+    return (
+        f'data:image/jpeg;base64,{image_base64}'
+        if add_data_prefix
+        else f'{image_base64}'
+    )
